Implementado/refatorado diversas coisas.

+++ Projeto_DDD-CQRS-ES/src/projeto/tcc/interfaceusuario/comandos/Comando.java
@@ -4,6 +4,6 @@ import java.util.UUID;
 
 public interface Comando {
 	UUID aggregateId();
-	Integer getVersion();
+	Long getVersion();
 
 }
