+++ Projeto_DDD-CQRS-ES/src/projeto/tcc/interfaceusuario/dto/FazerLoginDTO.java
@@ -9,7 +9,7 @@ public class FazerLoginDTO implements Serializable {
 	private static final long serialVersionUID = -7676555359721621926L;
 	private String login;
 	private String senha;
-	private Integer version;
+	private Long version;
 	
 	public FazerLoginDTO() {
 	}
@@ -30,11 +30,11 @@ public class FazerLoginDTO implements Serializable {
 		this.senha = senha;
 	}
 
-	public Integer getVersion() {
+	public Long getVersion() {
 		return version;
 	}
 
-	public void setVersion(Integer version) {
+	public void setVersion(Long version) {
 		this.version = version;
 	}
